move power gen func to power
