+++ nodes/sros/power.go
@@ -1,5 +1,10 @@
 package sros
 
+import (
+	"fmt"
+	"strings"
+)
+
 // SrosPower defines power supply configuration for a given node type.
 type SrosPower struct {
 	Modules any
@@ -40,4 +45,49 @@ var srosPowerConfig = map[string]SrosPower{
 		Modules: 10,
 		Shelves: 2,
 	},
+}
+
+func (n *sros) generatePowerConfig() string {
+	nodeType := strings.ToLower(n.Cfg.NodeType)
+	if _, ok := srosPowerConfig[nodeType]; !ok {
+		return ""
+	}
+
+	cfg := srosPowerConfig[nodeType]
+
+	shelves := 1
+	if s := cfg.Shelves; s != 0 {
+		shelves = s
+	}
+
+	modules := 0
+	switch m := cfg.Modules.(type) {
+	case map[string]int:
+		modules = m[defaultSrosPowerType]
+	case int:
+		modules = m
+	}
+
+	shelfType := fmt.Sprintf("ps-a%d-shelf-dc", modules)
+
+	var config strings.Builder
+
+	for s := 1; s <= shelves; s++ {
+		config.WriteString(
+			fmt.Sprintf(
+				"/configure chassis router chassis-number 1 power-shelf %d power-shelf-type %s\n",
+				s, shelfType))
+
+		for m := 1; m <= modules; m++ {
+			config.WriteString(
+				fmt.Sprintf(
+					"/configure chassis router chassis-number 1 power-shelf %d power-module %d power-module-type %s\n",
+					s,
+					m,
+					defaultSrosPowerModuleType,
+				))
+		}
+	}
+
+	return config.String()
 }
